+++ src/command_modules/azure-cli-lab/azure/cli/command_modules/lab/_help.py
@@ -135,13 +135,24 @@ helps['lab vm claim'] = """
             short-summary: Take ownership of an existing virtual machine in Azure DevTest Lab.
             parameters:
                 - name: --ids
-                  short-summary: One or more resource IDs (space delimited). If provided, no other arguments should be specified
+                  short-summary: One or more resource IDs (space delimited). If provided, no other arguments should be specified.
+                                 If --ids and --name both are omitted then provide --resource-group and --lab-name to claim any available VM.
                 - name: --resource-group -g
                   short-summary: Name of lab's resource group
                 - name: --lab-name
                   short-summary: Name of the Lab
                 - name: --name -n
                   short-summary: Name of the virtual machine
+            examples:
+                - name: Claim any available virtual machine in the lab.
+                  text: >
+                    az lab vm claim -g MyRG --lab-name MyLab
+                - name: Claim a specific virtual machine in the lab.
+                  text: >
+                    az lab vm claim -g MyRG --lab-name MyLab --name MyVM
+                - name: Claim multiple virtual machines in the lab using --ids.
+                  text: >
+                    az lab vm claim --ids /subscriptions/{SubID}/resourcegroups/{MyRG}/providers/microsoft.devtestlab/labs/{MyLab}/virtualmachines/{MyVM1} /subscriptions/{SubID}/resourcegroups/{MyRG}/providers/microsoft.devtestlab/labs/{MyLab}/virtualmachines/{MyVM2}
             """
 helps['lab custom-image'] = """
             type: group